NEW UPDATE: backend support... Kinda introduced a csv file that holds all the card information.
Now, board.js after it loads will read from the csv file and create the cards based on that data.
That way I do it is by using using the windows.fetch api to load the file and the library Papaparse to turn it into json.  Then the data loaded into the board.js state in the ComponentDidMount function.  During this time, the board will be loading and won't display the cards until the data has loaded.

So things I learned:
Using window.fetch API
Using Promises and how to make them asynchrous
Using Papaparse to parse the data
Learning the usage of ComponentDidMount.

Future topics:
Adding a deck, a graveyard and randomization.